Fix error where "watch" stopped watching after a Haxe error.
We had a problem because our loader was doing tricky things trying to run a haxe macro to figure out exactly which files to watch.

 Whenever there was an error, it didn't generate our file to figure out which files are used in the build, so webpack assumed no files were used, and stopped watching.  So you'd fix the error, and it wouldn't update.

Fix: instead of using the hacky macro to get a list of all files used, just use "addContextDependency" to track a whole directory for every "-cp" entry in your hxml file.

Nice!

As a result the "/macro" folder is now less magical and just a normal haxelib helper.  So I've renamed the folder to "haxelib".