Allow a configurable delay before emitting the fake server bundle.
This allows you to have something like

	nodemon actual-server.js

so that nodemon will restart your server when Haxe finishes building,
but 100ms later webpack will export `fake-server.bundle.js`, which if
you include it in your page:

	<script src="fake-server.bundle.js"></script>

it will trigger a page reload 100ms after the server was built,
which is hopefully enough time for it to restart.

Note: we're also adding a random string to the JS file to make
sure webpack notices the file is different and triggers a refresh.